requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
pydantic>=2.5.0
openai>=1.12.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
